requirements.txt
# TA超声图像年龄预测项目依赖

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
openpyxl>=3.0.9
Pillow>=9.0.0
opencv-python>=4.5.0  # 图像统计特征计算（analyze_error_samples.py）

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 工具
tqdm>=4.62.0
scikit-learn>=1.0.0

# 超参搜索
optuna>=3.0.0

# 可选：加速训练
# tensorboard>=2.8.0
